Rename the first utils file
